+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+v1.1
+----
+- Added core system files (23/02/2018)
+- Added libraries Bootstrap, jQuery, Fontawesome, CookieConsent (23/02/2018)
+
 v1.0
 ----
 - Creation of bundle (23/02/2018)